Allows splitting up embeddings generation
Embeddings generation is taking a long time, as I don't have a GPU and
have >1M documents. The code is now modularized so that I can run it in
chunks without losing progress (by passing different environment
variables) so that I can experiment a little with different models and
the like.